requirements.txt
Django>=4.2
djangorestframework>=3.14
ultralytics>=8.0
requests
PyYAML
